requirements.txt
asgiref==3.11.1
asttokens==3.0.1
colorama==0.4.6
comm==0.2.3
debugpy==1.8.20
decorator==5.2.1
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
executing==2.2.1
ipykernel==7.2.0
ipython==9.11.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.8.0
jupyter_core==5.9.1
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
packaging==26.0
parso==0.8.6
platformdirs==4.9.4
prompt_toolkit==3.0.52
psutil==7.2.2
pure_eval==0.2.3
Pygments==2.19.2
PyJWT==2.12.1
pyzmq==27.1.0
setuptools==82.0.1
sqlparse==0.5.5
stack-data==0.6.3
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
tzdata==2026.1
wcwidth==0.6.0
wheel==0.46.3
